Move removing of expired provider injections into GlobalPreferences
